requirements.txt
moviepy
librosa
google-generativeai
pexels-api-py
edge_tts
requests
nltk